adiciona testes de integração para operações CRUD de autores e ajusta rotas
