+++ jest.config.ts
@@ -4,7 +4,7 @@ const config: Config = {
     preset: 'ts-jest', // Define o preset do jest
     testEnvironment: 'node', // Define o ambiente de teste
     testSequencer: './custom-sequencer.js',  // Define o sequenciador de testes
-    maxWorkers: 1, // Define a quantidade de workers simultâneos
+    maxWorkers: 1, // Define a quantidade de workers simultâneos, se mais de 1 habilita o paralelismo
     // testPathPattern: '.*.spec.ts', // Define o padrão de busca dos arquivos de teste
     // globalSetup: Define o setup global de configuração
     // modulePathIgnorePatterns: [] Ignora os arquivos que estão no caminho
@@ -18,7 +18,7 @@ const config: Config = {
     reporters: [
         "default",
         ["./node_modules/jest-html-reporter", {
-            "pageTitle": "Test Report"
+            "pageTitle": "Relatório da Suíte de Testes"
         }]
     ],
     testLocationInResults: true
