+++ test/routes/author-route.ts
@@ -1,7 +1,7 @@
 export const authors = {
-    getListAuthors: '/api/v1/Authors',
-    createAuthors: '/api/v1/Authors',
-    getAuthorsById: '/api/v1/Authors/',
-    updateAuthors: '/api/v1/Authors/',
-    deleteAuthors:'/api/v1/Authors/'
+    getListAuthors: 'api/v1/Authors',
+    createAuthors: 'api/v1/Authors',
+    getAuthorsById: 'api/v1/Authors/',
+    updateAuthors: 'api/v1/Authors/',
+    deleteAuthors:'api/v1/Authors/'
 }
